# src/main/java/fr/epita/assistants/ping/presentation/api/request/user/UpdateUserRequest.java
package fr.epita.assistants.ping.presentation.api.request.user;

import fr.epita.assistants.ping.utils.HttpError;

public class UpdateUserRequest {
    public  String password;
    public String displayName;
    public String avatar;

}