# src/main/java/fr/epita/assistants/ping/presentation/api/response/user/LoginResponse.java
package fr.epita.assistants.ping.presentation.api.response.user;

import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
public class LoginResponse {
    public String token;
}